Rename entity, edit schema, add menu
