+++ custom_config_entity/src/CustomConfigEntityHtmlRouteProvider.php
@@ -7,12 +7,12 @@ use Drupal\Core\Entity\Routing\AdminHtmlRouteProvider;
 use Symfony\Component\Routing\Route;
 
 /**
- * Provides routes for Default entity entities.
+ * Provides routes for Custom config entity entities.
  *
  * @see Drupal\Core\Entity\Routing\AdminHtmlRouteProvider
  * @see Drupal\Core\Entity\Routing\DefaultHtmlRouteProvider
  */
-class DefaultEntityHtmlRouteProvider extends AdminHtmlRouteProvider {
+class CustomConfigEntityHtmlRouteProvider extends AdminHtmlRouteProvider {
 
   /**
    * {@inheritdoc}
