+++ custom_config_entity/src/CustomConfigEntityListBuilder.php
@@ -6,15 +6,15 @@ use Drupal\Core\Config\Entity\ConfigEntityListBuilder;
 use Drupal\Core\Entity\EntityInterface;
 
 /**
- * Provides a listing of Default entity entities.
+ * Provides a listing of Custom config entity entities.
  */
-class DefaultEntityListBuilder extends ConfigEntityListBuilder {
+class CustomConfigEntityListBuilder extends ConfigEntityListBuilder {
 
   /**
    * {@inheritdoc}
    */
   public function buildHeader() {
-    $header['label'] = $this->t('Default entity');
+    $header['label'] = $this->t('Custom config entity');
     $header['id'] = $this->t('Machine name');
     return $header + parent::buildHeader();
   }
